// generated/src/app/api-models/cap-entity-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapEntityWork {
  seqCcalcId: number;
  seqCapEntityWork: number;
  seqCapEligWork: number;
  seqProvId: number;
  seqVendId: number;
  capEntityCode: string;
  currentRetro: string;
  capEntityId: string;
  seqCapVendAddress: number;
  seqProvContract: number;
  maxEnrollLmt: number;
  excludeIncentive: string;
  accessProgramEligible: string;
}